fix(api): API構成を整理した
